Implement useGoogleLogin `auth-code` flow
This flow allows custom components to be used when triggering the Sign In With
Google box, but it requires additional validation of the returned "code" value,
which requires a client secret and a redirect uri. So it is slightly more
complicated, but more flexible.

Improve type hinting.

Update README.

Bump minimum Reflex to 0.6.6