ENG-7795: plumb through `scopes` parameter to get access_token
Allow interoperability with Google APIs by allowing the auth-code flow to
request scopes that enable it to get a real access token that can be used to
access google APIs on behalf of the authenticated user.

Updates the google_auth_demo to show how to request Drive scopes and update the
hidden app data in the user's Drive.